Move scrapeEngineTokudbStatus to collector package.
